chore: app setup - add docblocks to all PHP files - move validation handled in controller methods to request classes - rename validation classes to better match methods - add AbstractFormRequest and update form request classes to extend it - rename Controller to AbstractController - add AbstractModel and update User model to extend it - add base service configurations to AppServiceProvider - add class name to create* migrations and add constants for table names   - this way, we can use those constants in later migrations - limit table creation based on env config values - add AbstractSeeder - update favicon, logos, app name, and repo link - update .prettierrc and rerun formatter - add setup composer command - add architecture testing

+++ database/migrations/0001_01_01_000002_create_jobs_table.php
@@ -1,4 +1,10 @@
 <?php
+/**
+ * Create Jobs Table migration.
+ *
+ * @author Taylor Otwell <taylor@laravel.com>
+ * @author Nick Menke <git@nlmenke.net>
+ */
 
 declare(strict_types=1);
 
@@ -6,22 +12,45 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class() extends Migration
+/**
+ * Creates the database tables required for queued jobs.
+ *
+ * @since 0.0.0-framework introduced
+ * @since 0.0.0-vertebrae add classname and table constants
+ */
+final class CreateJobsTable extends Migration
 {
+    /**
+     * The main table used by the migration.
+     */
+    public const string TABLE = 'jobs';
+
+    /**
+     * The job batches table used by the migration.
+     */
+    public const string TABLE_BATCHES = 'job_batches';
+
+    /**
+     * The failed jobs table used by the migration.
+     */
+    public const string TABLE_FAILED = 'failed_jobs';
+
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::create('jobs', function (Blueprint $table): void {
-            $table->id();
-            $table->string('queue')->index();
-            $table->longText('payload');
-            $table->unsignedTinyInteger('attempts');
-            $table->unsignedInteger('reserved_at')->nullable();
-            $table->unsignedInteger('available_at');
-            $table->unsignedInteger('created_at');
-        });
+        if (config('queue.default') === 'database') {
+            Schema::create('jobs', function (Blueprint $table): void {
+                $table->id();
+                $table->string('queue')->index();
+                $table->longText('payload');
+                $table->unsignedTinyInteger('attempts');
+                $table->unsignedInteger('reserved_at')->nullable();
+                $table->unsignedInteger('available_at');
+                $table->unsignedInteger('created_at');
+            });
+        }
 
         Schema::create('job_batches', function (Blueprint $table): void {
             $table->string('id')->primary();
@@ -52,8 +81,11 @@ return new class() extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('jobs');
+        if (config('queue.default') === 'database') {
+            Schema::dropIfExists('jobs');
+        }
+
         Schema::dropIfExists('job_batches');
         Schema::dropIfExists('failed_jobs');
     }
-};
+}
